Add files via upload
This project is made in last semester of Bachelor Degree in Computer Engineering(Jan2022-April2022). Me and my friend worked on this project. It was a great experience doing this project because it helped us in learning team work, Web development and most importantly how to deal with a problem. Their was a time where we were having a viva and day before that our system crashed. It was not the main viva but like we need to show how much work we did. But due to mistake by us our system got crashed and need to buy everything again and then have to start over it as fresh. This teaches us that project should be done with extra care and properly as one wrong step and everything gets ruined. Making an application and writing a code of its working is the main part. It is quite challenging but that's what we are here for. If there won't be any difficulties than it not worth it working on it. So basically this project is quite interesting. We have made a very basic one but lot of modification can be done on it.